refactor(shared): 통합된 타입 정의로 의존성 정리
- ActionResult 타입을 shared/types로 이동
- 중복 및 불필요한 exports 제거
- 관련 파일들의 타입 정의 경로 업데이트
