+++ src/features/projects-create/api/index.ts
@@ -3,5 +3,4 @@ export {
   getProjects,
   checkProjectNameDuplicate,
   hashIdentifier,
-  type ActionResult,
 } from './server-action';
