+++ src/features/projects-create/api/server-action.ts
@@ -7,7 +7,6 @@ import type { CreateProjectDomain, ProjectDomain } from '@/entities';
 import { toProjectDto } from '@/entities';
 import { getDatabase, projects } from '@/shared/lib/db';
 import type { ActionResult } from '@/shared/types';
-export type { ActionResult } from '@/shared/types';
 import { eq } from 'drizzle-orm';
 import { v7 as uuidv7 } from 'uuid';
