+++ src/features/projects-create/index.ts
@@ -1,4 +1,3 @@
 export { createProject, getProjects, checkProjectNameDuplicate, hashIdentifier } from './api';
-export type { ActionResult } from './api';
 export { ProjectCreateForm } from './ui';
 export { createProjectMock, getProjectsMock, checkProjectNameDuplicateMock, resetMockDatabase, getMockDatabase } from './lib';
