+++ src/features/projects-create/lib/mock.ts
@@ -3,7 +3,7 @@
 import { revalidatePath } from 'next/cache';
 
 import type { CreateProjectDomain, ProjectDomain } from '@/entities';
-import type { ActionResult } from '@/features';
+import type { ActionResult } from '@/shared/types';
 import { hashIdentifier } from '@/features';
 import { v7 as uuidv7 } from 'uuid';
