+++ src/features/runs/api/index.ts
@@ -1,7 +1,4 @@
-export type { FetchedTestRun } from './get-test-runs';
 export { getTestRunsByProjectId } from './get-test-runs';
-export type { TestCaseRunDetail, SourceInfo, TestRunDetail } from './get-test-run-by-id';
 export { getTestRunById } from './get-test-run-by-id';
-export type { UpdateTestCaseRunInput, UpdateTestCaseRunResult } from './update-test-case-run';
 export { updateTestCaseRunStatus } from './update-test-case-run';
 export { testRunsQueryOptions, testRunByIdQueryOptions } from './query';
