+++ src/features/runs/index.ts
@@ -1,2 +1,2 @@
-export type { FetchedTestRun, TestCaseRunDetail, SourceInfo, TestRunDetail, UpdateTestCaseRunInput, UpdateTestCaseRunResult } from './api';
+export type { FetchedTestRun, TestCaseRunDetail, SourceInfo, TestRunDetail, UpdateTestCaseRunInput, UpdateTestCaseRunResult } from '@/entities/test-run';
 export { getTestRunsByProjectId, getTestRunById, updateTestCaseRunStatus, testRunsQueryOptions, testRunByIdQueryOptions } from './api';
